test: split tests by resource type. fix query time by limiting by timestamp
